+++ src/pages/SovereignTransfer/hooks/useSovereignTransferForm.ts
@@ -2,7 +2,7 @@ import BigNumber from 'bignumber.js';
 import { useFormik } from 'formik';
 import { useSelector } from 'react-redux';
 import { array, number, object, string } from 'yup';
-import { getSelectedTokenBalance } from 'helpers';
+import { addressIsErd, getSelectedTokenBalance } from 'helpers';
 import { useSendTransactions, useTokenOptions } from 'hooks';
 import { addressIsValid, useGetAccountInfo, useGetNetworkConfig } from 'lib';
 import { networkSelector } from 'redux/selectors';
@@ -49,7 +49,7 @@ export const useSovereignTransferForm = () => {
         .test(
           'addressIsValid',
           'Address is invalid',
-          (value) => !value || addressIsValid(value)
+          (value) => !value || addressIsValid(value) || addressIsErd(value)
         )
         .required('Receiver is required'),
       [SovereignTransferFormFieldsEnum.contract]: string()